refactor(config): extract yaml-loader.ts (R1)
Move YamlConfig, YAML_TEMPLATE, resolveEnvVars, loadYamlConfig, and
generateConfigTemplate byte-for-byte into smart-agent/yaml-loader.ts.
Re-export all moved symbols from config.ts barrel so public API is
unchanged. Add resolve-env-vars.test.ts covering substitution, fallback,
and deep nesting (7 tests).
